+++ stdlib/provider.go
@@ -121,6 +121,7 @@ func (*stdlibProvider) Functions(_ context.Context) []func() function.Function {
 		stringfunc.NewLastCharFunction,
 		stringfunc.NewLastIndexFunction,
 		stringfunc.NewSplitAfterFunction,
+		stringfunc.NewCompareStringFunction,
 	}
 }